refactor(app): flatten features into entity queries layer
Each entity now owns its full data contract:
- api.ts: raw Eden Treaty one-liners (no error handling, no parsing)
- queries.ts: all TanStack Query hooks (reads + mutations together)

Deletes the features/ directory entirely — auth, feed, pack, trip, and
user mutations move into their respective entity queries.ts files.
The package root re-exports everything from entities and shared.
