+++ packages/app/index.ts
@@ -3,20 +3,5 @@ export * from './src/entities/feed';
 export * from './src/entities/pack';
 export * from './src/entities/trip';
 export * from './src/entities/user';
-export { useLoginMutation } from './src/features/auth/login/hooks';
-export { useRegisterMutation } from './src/features/auth/register/hooks';
-export { useAddCommentMutation } from './src/features/feed/add-comment/hooks';
-export { useCreatePostMutation } from './src/features/feed/create-post/hooks';
-export { useTogglePostLikeMutation } from './src/features/feed/toggle-like/hooks';
-export { useAddPackItemMutation } from './src/features/pack/add-item/hooks';
-export { useCreatePackMutation } from './src/features/pack/create/hooks';
-export { useDeletePackMutation } from './src/features/pack/delete/hooks';
-export { useDeletePackItemMutation } from './src/features/pack/delete-item/hooks';
-export { useUpdatePackMutation } from './src/features/pack/update/hooks';
-export { useUpdatePackItemMutation } from './src/features/pack/update-item/hooks';
-export { useCreateTripMutation } from './src/features/trip/create/hooks';
-export { useDeleteTripMutation } from './src/features/trip/delete/hooks';
-export { useUpdateTripMutation } from './src/features/trip/update/hooks';
-export { useUpdateProfileMutation } from './src/features/user/update-profile/hooks';
 export * from './src/shared/api';
 export * from './src/shared/lib/weight';
