+++ packages/app/src/entities/catalog/queries.ts
@@ -1,16 +1,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { queryKeys, useApiClient } from '../../shared/api';
+import { type CatalogSortField, getCatalogItems } from './api';
 import { CatalogItemsResponseSchema } from './schema';
-
-type CatalogSortField =
-  | 'name'
-  | 'brand'
-  | 'category'
-  | 'price'
-  | 'ratingValue'
-  | 'createdAt'
-  | 'updatedAt'
-  | 'usage';
 
 interface UseCatalogItemsParams {
   query?: string;
@@ -29,14 +20,12 @@ export function useCatalogItemsInfinite({
   return useInfiniteQuery({
     queryKey: queryKeys.catalogInfinite(query, category),
     queryFn: async ({ pageParam = 1 }) => {
-      const { data, error } = await client.catalog.get({
-        query: {
-          page: pageParam as number,
-          limit,
-          ...(query ? { q: query } : {}),
-          ...(category ? { category } : {}),
-          ...(sort ? { sort } : {}),
-        },
+      const { data, error } = await getCatalogItems(client, {
+        page: pageParam as number,
+        limit,
+        ...(query ? { q: query } : {}),
+        ...(category ? { category } : {}),
+        ...(sort ? { sort } : {}),
       });
       if (error) throw new Error(`Failed to fetch catalog items: ${String(error)}`);
       const parseResult = CatalogItemsResponseSchema.safeParse(data);
