--- packages/app/src/entities/feed/hooks.ts
@@ -1,23 +1,0 @@
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { queryKeys, useApiClient } from '../../shared/api';
-
-export function useFeed() {
-  const client = useApiClient();
-  return useInfiniteQuery({
-    queryKey: queryKeys.feed(),
-    queryFn: async ({ pageParam = 1 }) => {
-      const { data, error } = await client.feed.get({
-        query: { page: pageParam as number, limit: 20 },
-      });
-      if (error) throw new Error(`Failed to fetch feed: ${String(error)}`);
-      return data;
-    },
-    getNextPageParam: (lastPage) => {
-      if (lastPage && lastPage.page < lastPage.totalPages) {
-        return lastPage.page + 1;
-      }
-      return undefined;
-    },
-    initialPageParam: 1,
-  });
-}
