+++ packages/app/src/entities/feed/index.ts
@@ -1,4 +1,8 @@
-export { fetchFeed } from './api';
-export { useFeed } from './hooks';
+export {
+  useAddCommentMutation,
+  useCreatePostMutation,
+  useFeed,
+  useTogglePostLikeMutation,
+} from './queries';
 export { CommentSchema, FeedResponseSchema, PostAuthorSchema, PostSchema } from './schema';
 export type { Comment, FeedResponse, Post } from './types';
