+++ packages/app/src/entities/pack/api.ts
@@ -1,14 +1,93 @@
 import type { ApiClient } from '../../shared/api';
-import { PackListResponseSchema, PackSchema } from './schema';
 
-export async function fetchPacks(client: ApiClient) {
-  const { data, error } = await client.packs.get({ query: { includePublic: 0 } });
-  if (error) throw new Error('Failed to fetch packs');
-  return PackListResponseSchema.parse(data);
-}
+export const getPacks = (client: ApiClient) => client.packs.get({ query: { includePublic: 0 } });
 
-export async function fetchPack(client: ApiClient, packId: string) {
-  const { data, error } = await client.packs({ packId }).get();
-  if (error) throw new Error('Failed to fetch pack');
-  return PackSchema.parse(data);
-}
+export const getPack = (client: ApiClient, packId: string) => client.packs({ packId }).get();
+
+export const createPack = (
+  client: ApiClient,
+  body: {
+    id: string;
+    name: string;
+    description?: string;
+    category?: string;
+    isPublic: boolean;
+    image?: string | null;
+    tags?: string[];
+    localCreatedAt: string;
+    localUpdatedAt: string;
+  },
+) => client.packs.post(body);
+
+export const updatePack = (
+  client: ApiClient,
+  {
+    packId,
+    body,
+  }: {
+    packId: string;
+    body: {
+      name?: string;
+      description?: string;
+      category?: string;
+      isPublic?: boolean;
+      image?: string | null;
+      tags?: string[];
+      deleted?: boolean;
+      localUpdatedAt?: string;
+    };
+  },
+) => client.packs({ packId }).put(body);
+
+export const deletePack = (client: ApiClient, packId: string) => client.packs({ packId }).delete();
+
+export const addPackItem = (
+  client: ApiClient,
+  {
+    packId,
+    body,
+  }: {
+    packId: string;
+    body: {
+      id: string;
+      name: string;
+      description?: string;
+      weight: number;
+      weightUnit: 'g' | 'oz' | 'kg' | 'lb';
+      quantity: number;
+      consumable: boolean;
+      worn: boolean;
+      category?: string;
+      image?: string | null;
+      notes?: string | null;
+      catalogItemId?: number | null;
+    };
+  },
+) => client.packs({ packId }).items.post(body);
+
+export const updatePackItem = (
+  client: ApiClient,
+  {
+    itemId,
+    body,
+  }: {
+    itemId: string;
+    body: {
+      name?: string;
+      description?: string;
+      weight?: number;
+      weightUnit?: 'g' | 'oz' | 'kg' | 'lb';
+      quantity?: number;
+      category?: string;
+      consumable?: boolean;
+      worn?: boolean;
+      image?: string | null;
+      notes?: string | null;
+      catalogItemId?: number | null;
+      deleted?: boolean;
+    };
+  },
+) => client.packs.items({ itemId }).patch(body);
+
+export const deletePackItem = (client: ApiClient, itemId: string) =>
+  client.packs.items({ itemId }).delete();
