+++ packages/app/src/entities/pack/index.ts
@@ -1,5 +1,13 @@
-export { fetchPack, fetchPacks } from './api';
-export { usePack, usePacks } from './hooks';
+export {
+  useAddPackItemMutation,
+  useCreatePackMutation,
+  useDeletePackItemMutation,
+  useDeletePackMutation,
+  usePack,
+  usePacks,
+  useUpdatePackItemMutation,
+  useUpdatePackMutation,
+} from './queries';
 export {
   PackItemSchema,
   PackListResponseSchema,
