+++ packages/app/src/entities/trip/api.ts
@@ -1,14 +1,43 @@
 import type { ApiClient } from '../../shared/api';
-import { TripSchema } from './schema';
 
-export async function fetchTrips(client: ApiClient) {
-  const { data, error } = await client.trips.get();
-  if (error) throw new Error('Failed to fetch trips');
-  return TripSchema.array().parse(data);
-}
+export const getTrips = (client: ApiClient) => client.trips.get();
 
-export async function fetchTrip(client: ApiClient, tripId: string) {
-  const { data, error } = await client.trips({ tripId }).get();
-  if (error) throw new Error('Failed to fetch trip');
-  return TripSchema.parse(data);
-}
+export const getTrip = (client: ApiClient, tripId: string) => client.trips({ tripId }).get();
+
+export const createTrip = (
+  client: ApiClient,
+  body: {
+    id: string;
+    name: string;
+    description?: string | null;
+    notes?: string | null;
+    location?: { latitude: number; longitude: number; name?: string } | null;
+    startDate?: string | null;
+    endDate?: string | null;
+    packId?: string | null;
+    localCreatedAt: string;
+    localUpdatedAt: string;
+  },
+) => client.trips.post(body);
+
+export const updateTrip = (
+  client: ApiClient,
+  {
+    tripId,
+    body,
+  }: {
+    tripId: string;
+    body: {
+      name?: string;
+      description?: string | null;
+      notes?: string | null;
+      location?: { latitude: number; longitude: number; name?: string } | null;
+      startDate?: string | null;
+      endDate?: string | null;
+      packId?: string | null;
+      localUpdatedAt?: string;
+    };
+  },
+) => client.trips({ tripId }).put(body);
+
+export const deleteTrip = (client: ApiClient, tripId: string) => client.trips({ tripId }).delete();
