+++ packages/app/src/entities/trip/index.ts
@@ -1,4 +1,9 @@
-export { fetchTrip, fetchTrips } from './api';
-export { useTrip, useTrips } from './hooks';
+export {
+  useCreateTripMutation,
+  useDeleteTripMutation,
+  useTrip,
+  useTrips,
+  useUpdateTripMutation,
+} from './queries';
 export { TripLocationSchema, TripSchema } from './schema';
 export type { Trip, TripLocation } from './types';
