+++ packages/app/src/entities/user/index.ts
@@ -1,4 +1,9 @@
-export { fetchCurrentUser, fetchUserProfile } from './api';
-export { useCurrentUser, useUserProfile } from './hooks';
+export {
+  useCurrentUser,
+  useLoginMutation,
+  useRegisterMutation,
+  useUpdateProfileMutation,
+  useUserProfile,
+} from './queries';
 export { UserProfileSchema, UserSchema } from './schema';
 export type { User, UserProfile } from './types';
