--- packages/app/src/features/pack/create/hooks.ts
@@ -1,32 +1,0 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useApiClient } from '../../../shared/api';
-import { queryKeys } from '../../../shared/api/query-keys';
-
-interface CreatePackInput {
-  name: string;
-  description?: string;
-  category?: string;
-  isPublic?: boolean;
-  image?: string | null;
-  tags?: string[];
-}
-
-export function useCreatePackMutation() {
-  const client = useApiClient();
-  const qc = useQueryClient();
-  return useMutation({
-    mutationFn: async (input: CreatePackInput) => {
-      const now = new Date().toISOString();
-      const { data, error } = await client.packs.post({
-        ...input,
-        id: crypto.randomUUID(),
-        isPublic: input.isPublic ?? false,
-        localCreatedAt: now,
-        localUpdatedAt: now,
-      });
-      if (error) throw new Error('Failed to create pack');
-      return data;
-    },
-    onSuccess: () => qc.invalidateQueries({ queryKey: queryKeys.packs() }),
-  });
-}
